Fix ga and gA commands for files bigger than 4 GiB
For 32-bit operating systems.

Thanks to filterfalse.

+++ src/utils/fs.c
@@ -371,7 +371,7 @@ get_file_size(const char *path)
 	struct stat st;
 	if(lstat(path, &st) == 0)
 	{
-		return (size_t)st.st_size;
+		return (uint64_t)st.st_size;
 	}
 	return 0;
 #else
